// frontend/src/generated/model/account-info-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * All information of a specific account
 */
export interface AccountInfoDto { 
    /**
     * Username of the account to be created.
     */
    username: string;
    /**
     * Email of the account to be created.
     */
    email: string;
    /**
     * First name of the user that creates this account.
     */
    firstName: string;
    /**
     * Last name of the user that creates this account.\"
     */
    lastName: string;
    /**
     * Indicates if the account is an admin.
     */
    isGlobalAdmin: boolean;
}